fix: Minor fix
